+++ src/gui/widgets/dialogs/trace_dialog.hpp
@@ -32,6 +32,7 @@ private slots:
     void auto_colors();
     void zoom_preview(qreal percent);
     void show_help();
+    void preview_slide(int percent);
 
 private:
     class Private;